+++ packages/utilities/psammead-storybook-helpers/src/index.js
@@ -1,5 +1,5 @@
-import inputProvider from './input-provider';
-import { dirDecorator } from './decorators';
+import withServicesKnob from './withServicesKnob';
+import buildRTLSubstories from './buildRTLSubstories';
 import LANGUAGE_VARIANTS from './text-variants';
 
-export { inputProvider, LANGUAGE_VARIANTS, dirDecorator };
+export { LANGUAGE_VARIANTS, withServicesKnob, buildRTLSubstories };